Explicit void type
Void type is required so that a procedure can be considered as a type. It also makes 'proc' keyword redundant. A 'func' without any return type is considered a procedure as well as when Void type is given - it is not perfect, but removing any option does n't either seem like a good idea.