Functions without final 'return'
Void-returning functions do not need to have a return instruction. All other functions should return a value on every path, and if they don't, the result is undefined and the program may break. In the future some warnings might prove useful, as well as displaying explicit runtime errors in such situations.