Default arguments in other modules
For default arguments (or generally global variables) to work in different modules, it is easiest to compile all code to one LLVM file (but the source code files are not simply concatenated now). Some mechanism to compile only used functions would be nice, so the generated code is more readable, but it would make everything more complicated.